Add: Images, Modify function, Refactoring some files
