+++ src/Pages/Store/Home/Component/Jumbotron/Jumbotron.scss
@@ -2,16 +2,13 @@
 @import url("https://spoqa.github.io/spoqa-han-sans/ko-KR/");
 
 .jumbotronContainer {
-  width: 1280px;
+  width: 1240px;
   height: 675px;
   margin: 138px auto 100px auto;
-  // margin-top: 138px;
-  // margin-bottom: 100px;
-  // object-fit: fill;
   font-family: "Spoqa Han Sans", "Spoqa Han Sans JP", "Sans-serif";
   .item {
     .jumboImg {
-      width: 1280px;
+      width: 1240px;
       height: 675px;
       object-fit: fill;
     }
