--- src/Pages/Store/Home/Component/Navbar/Navbar.scss
@@ -1,294 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  box-sizing: border-box;
-  // font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-  //   monospace;
-}
-
-$green: #4adaaf;
-$white: #ffffff;
-$black: #26282b;
-$ootdBorderColor: #e8ebeb;
-$gray: #9ea4aa;
-$blue: #3399ff;
-$storeBorderColor: #f0f1f4;
-
-.navBarWrapper {
-  position: fixed;
-  width: 100%;
-  z-index: 1;
-  top: 0;
-  transition-property: box-shadow;
-  transition-duration: 0.25s;
-  transition-timing-function: ease-in-out;
-  background: $white;
-  box-shadow: rgba(27, 29, 31, 0.12) 0px 2px 10px;
-  color: $black;
-  font-family: "Helvetica", "Arial", "Apple SD Gothic Neo", "Nanum Gothic",
-    "Malgun Gothic", "DotumChe", "Dotum", sans-serif;
-  line-height: 1.3;
-  background-color: white;
-
-  .topNavBar {
-    display: flex;
-    align-items: center;
-    z-index: 2;
-    background: $white;
-
-    .leftNavBar {
-      display: flex;
-      -webkit-box-align: center;
-      align-items: center;
-      padding-left: 80px;
-
-      .navLogo {
-        outline: none;
-        cursor: pointer;
-        text-decoration: underline;
-
-        img {
-          width: 44px;
-          height: 44px;
-        }
-      }
-
-      .leftMenuWrapper {
-        flex-grow: 0;
-        flex-shrink: 0;
-        flex-basis: auto;
-      }
-
-      .leftMenuList {
-        list-style: none;
-        text-align: center;
-        margin-block-start: 1em;
-        margin-block-end: 1em;
-        margin-inline-start: 0px;
-        margin-inline-end: 0px;
-        padding-inline-start: 10px;
-      }
-
-      .leftMenu {
-        display: inline-block;
-        vertical-align: middle;
-        text-align: center;
-        font-size: 14px;
-        font-weight: 400;
-        letter-spacing: -0.4px;
-        line-height: 1.43;
-      }
-
-      .leftMenuOotd {
-        display: block;
-        text-decoration: none;
-        font-weight: 700;
-        color: rgb(27, 29, 31);
-        cursor: pointer;
-        border: 0px;
-        background: transparent;
-        padding: 10px 12px;
-        outline: none;
-        font-size: 14px;
-        line-height: 1.43;
-        cursor: pointer;
-        transform: translateZ(0);
-        transition: transform 0.25s ease-out;
-        &:hover {
-          transform: scale(1.05);
-        }
-      }
-
-      .leftMenuOotdAct {
-        display: block;
-        text-decoration: none;
-        font-weight: 700;
-        color: rgb(27, 29, 31);
-        cursor: pointer;
-        border-bottom: 1px solid black;
-        background: transparent;
-        padding: 10px 12px;
-        outline: none;
-        font-size: 14px;
-        line-height: 1.43;
-        cursor: pointer;
-        transform: translateZ(0);
-        transition: transform 0.25s ease-out;
-        &:hover {
-          transform: scale(1.05);
-        }
-      }
-
-      .leftMenuStoreAct {
-        display: block;
-        text-decoration: none;
-        font-weight: 700;
-        color: rgb(27, 29, 31);
-        cursor: pointer;
-        border-bottom: 1px solid black;
-        background: transparent;
-        padding: 10px 12px;
-        outline: none;
-        font-size: 14px;
-        line-height: 1.43;
-        cursor: pointer;
-        transform: translateZ(0);
-        transition: transform 0.25s ease-out;
-        &:hover {
-          transform: scale(1.05);
-        }
-      }
-
-      .leftMenuStore {
-        display: block;
-        text-decoration: none;
-        font-weight: 700;
-        color: rgb(27, 29, 31);
-        cursor: pointer;
-        background: transparent;
-        padding: 10px 12px;
-        outline: none;
-        font-size: 14px;
-        line-height: 1.43;
-        cursor: pointer;
-        transform: translateZ(0);
-        transition: transform 0.25s ease-out;
-        &:hover {
-          transform: scale(1.05);
-        }
-      }
-      span {
-        color: rgb(27, 29, 31);
-      }
-    }
-
-    .rightNavBar {
-      display: flex;
-      position: absolute;
-      -webkit-box-align: center;
-      align-items: center;
-      right: 80px;
-
-      .rightMenuWrapper {
-        display: flex;
-        align-items: center;
-        margin-left: 4px;
-
-        .rightMenuList {
-          list-style: none;
-          text-align: center;
-          margin: 0px;
-          padding: 0px;
-          margin-block-start: 1em;
-          margin-block-end: 1em;
-          margin-inline-start: 0px;
-          margin-inline-end: 0px;
-          padding-inline-start: 40px;
-
-          li {
-            display: inline-block;
-            vertical-align: middle;
-            font-weight: 700;
-            letter-spacing: -0.4px;
-            padding-right: 18px;
-
-            svg {
-              width: 28px;
-              height: 28px;
-              cursor: pointer;
-            }
-
-            img {
-              width: 28px;
-              height: 28px;
-              cursor: pointer;
-              border: 0px solid black;
-              border-radius: 50%;
-            }
-          }
-        }
-      }
-    }
-
-    .searchBar {
-      margin: 0 auto;
-      transform: translateX(-20%);
-      width: 734.797px;
-      height: 40px;
-      z-index: 102;
-      font-size: 14px;
-      font-weight: 400;
-      letter-spacing: -0.4px;
-      line-height: 1.43;
-      color: rgb(27, 29, 31);
-      background-color: rgb(247, 248, 249);
-      border-radius: 22px;
-      border-width: 1px;
-      border-style: solid;
-      border-color: rgb(232, 235, 237);
-      border-image: initial;
-      transition: background-color 0.25s ease-in-out 0s, color 0.25s 0s;
-
-      input {
-        label {
-          svg {
-            margin-right: 12px;
-            width: 20px;
-            height: 20px;
-          }
-        }
-        appearance: none;
-        width: 100%;
-        height: 100%;
-        text-align: center;
-        transition: color 0.25s ease-in-out 0s;
-        outline: none;
-        background-image: url("http://www.w3.org/2000/svg");
-        background-color: transparent;
-        border: 0px;
-        -webkit-tap-highlight-color: transparent;
-        font-size: 14px;
-        font-weight: 400;
-        letter-spacing: -0.4px;
-        line-height: 1.43;
-        color: #26282b;
-        font-family: "Helvetica", "Arial", "Apple SD Gothic Neo", "Nanum Gothic",
-          "Malgun Gothic", "DotumChe", "Dotum", sans-serif;
-      }
-    }
-  }
-
-  .bottomNavBar {
-    display: flex;
-    height: 64px;
-    justify-content: center;
-    align-items: center;
-    z-index: 0;
-
-    .categories {
-      list-style: none;
-      text-align: left;
-      font-size: 16px;
-      font-weight: 700;
-      letter-spacing: -0.4px;
-      line-height: 1.31;
-    }
-
-    li {
-      display: inline-block;
-      vertical-align: middle;
-      padding: 10px 12px;
-      cursor: pointer;
-      transform: translateZ(0);
-      transition: transform 0.25s ease-out;
-      &:hover {
-        transform: scale(1.05);
-      }
-    }
-  }
-}
