+++ src/Pages/Store/Home/Component/PopularBrand/PopularBrand.js
@@ -2,10 +2,34 @@ import React, { Component } from "react";
 import "./PopularBrand.scss";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
+// import "slick-carousel/slick/slick-theme.css";
 
 class PopularBrand extends Component {
+  constructor() {
+    super();
+    this.state = {
+      brandList: [],
+    };
+  }
+
+  getBrandList = () => {
+    fetch("http://localhost:3000/data/brandItem.json", {
+      method: "GET",
+    })
+      .then((res) => res.json())
+      .then((res) => {
+        this.setState({
+          brandList: res.item,
+        });
+      });
+  };
+  componentDidMount() {
+    this.getBrandList();
+  }
+
   render() {
+    const { brandList } = this.state;
+
     const settings = {
       dots: false,
       infinite: true,
@@ -17,24 +41,21 @@ class PopularBrand extends Component {
           breakpoint: 1024,
           settings: {
             slidesToShow: 3,
-            slidesToScroll: 1.5,
-            infinite: true,
-            dots: true,
+            slidesToScroll: 1,
           },
         },
         {
           breakpoint: 600,
           settings: {
             slidesToShow: 3,
-            slidesToScroll: 1.5,
-            initialSlide: 2,
+            slidesToScroll: 1,
           },
         },
         {
           breakpoint: 480,
           settings: {
             slidesToShow: 3,
-            slidesToScroll: 1.5,
+            slidesToScroll: 1,
           },
         },
       ],
@@ -45,9 +66,9 @@ class PopularBrand extends Component {
         <div className="popularBrandContainer">
           <header>
             <div className="title">
-              <h2>인기 상품</h2>
+              <h2>인기 브랜드</h2>
             </div>
-            {/* <div className="carouselBtn">
+            <div className="carouselBtn">
               <img
                 className="prevBtn"
                 src="images/icon/prev1.png"
@@ -58,121 +79,36 @@ class PopularBrand extends Component {
                 src="images/icon/next1.png"
                 alt="우측으로 넘기는 버튼"
               ></img>
-            </div> */}
+            </div>
           </header>
           <div className="carousel">
             <Slider {...settings}>
-              <div className="brandItem">
-                <div className="brandImgBox">
-                  <img
-                    className="brandItemImg"
-                    src="images/brand/brand0.jpg"
-                    alt="브랜드"
-                  ></img>
-                  <div className="brandItemDescBox">
-                    <div className="brandItemDesc">
-                      <span className="brandName">모르모트</span>
-                      <span className="brandProductCnt">( 329 )</span>
+              {brandList.map((brand) => {
+                return (
+                  <div className="brandItem">
+                    <div className="brandImgBox">
+                      <img
+                        className="brandItemImg"
+                        src={brand.src}
+                        alt={brand.alt}
+                      />
+                    </div>
+                    <div className="brandItemDescBox">
+                      <div className="brandItemDesc">
+                        <span className="brandName">{brand.brandName}</span>
+                        <span className="brandProductCnt">
+                          ({brand.brandProductCnt})
+                        </span>
+                      </div>
                     </div>
                   </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand1.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand2.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand3.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand4.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand5.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand6.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
-              <div className="brandItem">
-                <img
-                  className="brandItemImg"
-                  src="images/brand/brand7.jpg"
-                  alt="브랜드"
-                ></img>
-                <div className="brandItemDescBox">
-                  <div className="brandItemDesc">
-                    <span className="brandName">모르모트</span>
-                    <span className="brandProductCnt">( 329 )</span>
-                  </div>
-                </div>
-              </div>
+                );
+              })}
             </Slider>
           </div>
           <footer>
             <button>
-              <span>브랜드 더 보기 ></span>
+              <span>브랜드 더 보기 `{'>'}`</span>
             </button>
           </footer>
         </div>
