+++ src/Pages/Store/Home/Component/PopularProduct/PopularProduct.scss
@@ -2,20 +2,16 @@
 @import url("https://spoqa.github.io/spoqa-han-sans/ko-KR/");
 
 .popularContainer {
-  // display: flex;
-  // justify-content: center;
   font-family: "Spoqa Han Sans", "Spoqa Han Sans JP", "Sans-serif";
-  width: 1280px;
+  width: 1240px;
   height: 600px;
   margin: 100px auto;
-
-  // border: 3px solid red;
   .title {
     display: flex;
     justify-content: center;
     width: 100%;
     height: 50px;
-    // border: 1px solid red;
+
     h2 {
       font-size: 46px;
       margin-top: 0px;
@@ -50,19 +46,22 @@
     display: flex;
     width: 100%;
     height: 450px;
-    // border: 1px solid red;
+
     .card {
-      width: 320px;
+      width: 310px;
       height: 400px;
-      // border: 1px solid red;
+
       .productImgBox {
-        width: 320px;
-        height: 320px;
+        display: flex;
+        justify-content: center;
+        width: 310px;
+        height: 310px;
         overflow: hidden;
         .productImg {
-          width: 90%;
-          height: 320px;
-          padding: 0 5px;
+          width: 98%;
+          height: 100%;
+          // padding: 0 5px;
+          border-radius: 15px;
           &:hover {
             transform: scale(1.1);
             transition: transform 0.5s ease;
@@ -75,7 +74,6 @@
         .brandName {
           font-weight: 700;
           font-size: 18px;
-          // padding: 20px 0 20px 0;
         }
         .productName {
           width: 280px;
@@ -119,7 +117,6 @@
       border-radius: 60px;
       border: none;
       &:hover {
-        // background-color: white;
         opacity: 0.5;
         cursor: pointer;
       }
