+++ src/Pages/Store/Home/Home.js
@@ -1,8 +1,9 @@
 import React, { Component } from "react";
-import Navbar from "./Component/Navbar/Navbar";
+import Navbar from "../../../Component/Navbar/NavbarStore/Navbar";
 import Jumbotron from "./Component/Jumbotron/Jumbotron";
 import PopularProduct from "./Component/PopularProduct/PopularProduct";
 import PopularBrand from "./Component/PopularBrand/PopularBrand";
+import Trend from "./Component/Trend/Trend";
 
 class Home extends Component {
   render() {
@@ -12,6 +13,7 @@ class Home extends Component {
         <Jumbotron />
         <PopularProduct />
         <PopularBrand />
+        <Trend />
       </>
     );
   }
