+++ src/Styles/common.scss
@@ -1,3 +1,6 @@
+@import url("https://fonts.googleapis.com/css2?family=Nanum+Gothic:wght@400;700;800&display=swap");
+@import url("https://spoqa.github.io/spoqa-han-sans/ko-KR/");
+
 $green: #4adaaf;
 $white: #ffffff;
 $black: #26282b;
