+++ src/Styles/reset.scss
@@ -128,3 +128,16 @@ ol,
 li {
   list-style: none;
 }
+
+* {
+  margin: 0;
+  padding: 0;
+  font-family: "Spoqa Han Sans", "Spoqa Han Sans JP", "Sans-serif",
+    -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu",
+    "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  box-sizing: border-box;
+  // font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+  //   monospace;
+}
