+++ src/index.js
@@ -2,5 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import Routes from "./Routes";
 import "./Styles/reset.scss";
+import "./Styles/common.scss";
 
 ReactDOM.render(<Routes />, document.getElementById("root"));
